Crate noodles_csi

noodles-csi handles the reading and writing of the coordinate-sorted index (CSI) format.

Re-exports

pub use self::binning_index::BinningIndex;

pub use self::index::Index;

Modules

binning_index

Binning index utilities.

index

Coordinate-sorted index and fields.

Structs

Reader

A CSI reader.

Writer

A CSI writer.

Traits

BinningIndexReferenceSequence

An extension that adds methods to binning index reference sequence types.

Functions

read

Reads the entire contents of a coordinate-sorted index (CSI).

write

Writes a coordinate-sorted index (CSI) to a file.