Trait noodles_csi::binning_index::BinningIndex

pub trait BinningIndex {
    type ReferenceSequence: ReferenceSequenceExt;

    fn reference_sequences(&self) -> &[Self::ReferenceSequence];
    fn unplaced_unmapped_record_count(&self) -> Option<u64>;
    fn query<I>(
        &self, 
        reference_sequence_id: usize, 
        interval: I
    ) -> Result<Vec<Chunk>>
    where
        I: Into<Interval>;

    fn first_record_in_last_linear_bin_start_position(
        &self
    ) -> Option<VirtualPosition> { ... }
}

A binning index.

Required Associated Types

type ReferenceSequence: ReferenceSequenceExt

The returned output indexed reference sequence.

Required Methods

fn reference_sequences(&self) -> &[Self::ReferenceSequence]

Returns a list of indexed reference sequences.

fn unplaced_unmapped_record_count(&self) -> Option<u64>

Returns the number of unplaced, unmapped records in the associated file.

fn query<I>(
&self,
reference_sequence_id: usize,
interval: I
) -> Result<Vec<Chunk>> where
I: Into<Interval>,

Returns the chunks that overlap with the given region.

Provided Methods

fn first_record_in_last_linear_bin_start_position(
&self
) -> Option<VirtualPosition>

Returns the start position of the first record in the last linear bin.

This is the closest position to the unplaced, unmapped records, if any, that is available in an index.