Struct noodles_csi::index::Index

pub struct Index { /* private fields */ }

A coordinate-sorted index (CSI).

Implementations

impl Index

pub fn builder() -> Builder

Returns a builder to create an index from each of its fields.

Examples

use noodles_csi as csi;
let builder = csi::Index::builder();

pub fn min_shift(&self) -> u8

Returns the number of bits for the minimum interval.

Examples

use noodles_csi as csi;
let index = csi::Index::default();
assert_eq!(index.min_shift(), 14);

pub fn depth(&self) -> u8

Returns the depth of the binning index.

Examples

use noodles_csi as csi;
let index = csi::Index::default();
assert_eq!(index.depth(), 5);

pub fn aux(&self) -> &[u8]

Returns the auxiliary data.

Examples

use noodles_csi as csi;
let index = csi::Index::default();
assert!(index.aux().is_empty());

pub fn unmapped_read_count(&self) -> Option<u64>

👎 Deprecated since 0.2.0:
Use unplaced_unmapped_record_count instead.

Returns the number of unmapped records in the associated file.

Examples

use noodles_csi as csi;
let index = csi::Index::default();
assert!(index.unmapped_read_count().is_none());